backend/app/requirements.txt
flask
sklearn
pandas
numpy
requests
pillow
pytesseract
joblib
